move fn back to analytics file and fix up doc
